requirements.txt
Django==2.2.10
django-extra-fields==2.0.5
djangorestframework==3.11.0
pkg-resources==0.0.0
psycopg2==2.8.5
pytz==2020.1
sqlparse==0.3.1
